Compiler and placeholders for eBook
This generates both epub and pdf automatically from html.

Html will be a central format because it can also be published on github
pages.

But very likely I'd like a more text-based source format, so I'll
investigate the option to create these html files from markdown. I'll
need some scripting to update content table and so.

This is a nice first stepm though :-)